Use BraveVPNRegionDataManager for region data handling
BraveVPNOSConnectionAPIBase owns this manager.
